+++ src/screens/main/GreetingScreen.tsx
@@ -1,21 +1,21 @@
 import React from 'react'
 
 import {ThemeProvider} from 'styled-components/native'
-import theme from '../theme'
-import {View, Text, Button, Image} from '../base'
-import {Header} from '../components/hellostyledworld'
+import theme from '../../theme'
+import {View, Text, Button, Image} from '../../base'
+import {Header} from '../../components/hellostyledworld'
 import {Alert} from 'react-native'
 
-export default function IndependenceScreen({navigation}) {
+export default function GreetingScreen({navigation}) {
   return (
     <ThemeProvider theme={theme}>
-      <Header text="Independence" />
+      <Header text="Greetings" />
       <View flex={3} backgroundColor="lightGrey">
         <Text fontSize={36} paddingY={20} paddingX={20}>
           Training:
         </Text>
-        <Text>- no whining at night</Text>
-        <Text>- no whining when I'm not in the same room as her</Text>
+        <Text>- Dog to Dog greeting nicely</Text>
+        <Text>- Dog to person greeting nicely</Text>
       </View>
       <View flex={0.5} marginY={20}>
         <Button
