+++ src/types.ts
@@ -1,0 +1,6 @@
+// import {ParamListBase} from '@react-navigation/core'
+
+// // https://github.com/react-navigation/react-navigation/issues/6931
+// export type SubNavigator<T extends ParamListBase> = {
+//   [K in keyof T]: {screen: K; params?: T[K]}
+// }[keyof T]
